rails does not support double data type. had to force a new model with a decimal data type instead then migrated

+++ db/migrate/20190319051046_create_products.rb
@@ -1,11 +1,11 @@
 class CreateProducts < ActiveRecord::Migration[5.2]
   def change
     create_table :products do |t|
-      t.int :UserId
+      t.integer :UserId
       t.string :Name
       t.string :Description
-      t.int :Quantity
-      t.double :Price
+      t.integer :Quantity
+      t.decimal :Price
 
       t.timestamps
     end
